Generalized addon extraction code and moved it to separate module
