+++ lib/tasks/extract-component.js
@@ -1,66 +1,21 @@
 /*jshint node:true*/
 'use strict';
 
-var PLACEHOLDER = '<%= componentName %>';
-
-var Promise = require('../ext/promise');
-var fs = require('fs-extra');
 var path = require('path');
+var Promise = require('../ext/promise');
 var ui = require('../ui');
+var extractAddon = require('../utils/extract-addon');
 
-
-var ensureDir = Promise.denodeify(fs.ensureDir);
-var copy = Promise.denodeify(fs.copy);
-var readFile = Promise.denodeify(fs.readFile);
-var writeFile = Promise.denodeify(fs.writeFile);
-
-function createAddonFolder(name) {
-  return ensureDir(path.join('addon', name));
+function getFilesWithPlaceholdersFromName(name) {
+  return [
+    path.join('addon', name, 'index.js'),
+    path.join('addon', name, 'package.json'),
+    path.join('addon', name, 'style.css')
+  ];
 }
 
-function copyBlueprintFiles(name, blueprintName) {
-  var blueprintFolderRelativePath = "../../blueprints";
-  var blueprintPath = path.resolve(__dirname, blueprintFolderRelativePath, blueprintName);
-
-  var srcPath = path.join(blueprintPath, 'files');
-  var destPath = path.join('addon', name);
-
-  return copy(srcPath, destPath);
-}
-
-function replaceInFile(file, placeholder, value) {
-  return readFile(file, { encoding: 'utf-8' }).then(function(data) {
-    var modifiedData = data.replace(placeholder, value);
-    return writeFile(file, modifiedData);
-  });
-}
-
-function replaceInFiles(files, placeholder, value) {
-  return Promise.map(files, function(file) {
-    return replaceInFile(file, placeholder, value);
-  });
-}
-
-function insertAddonName(placeholder, name) {
-  var indexJsFile = path.join('addon', name, 'index.js');
-  var packageJsonFile = path.join('addon', name, 'package.json');
-  var styleCssFile = path.join('addon', name, 'style.css');
-
-  var files = [indexJsFile, packageJsonFile, styleCssFile];
-
-  return replaceInFiles(files, placeholder, name).then(function() {
-    return Promise.resolve(name);
-  });
-}
-
-function copyFileIfExists(oldFile, newFile) {
-  return new Promise(function(resolve) {
-    copy(oldFile, newFile, { clobber: true }).finally(resolve);
-  });
-}
-
-function overwriteFiles(name) {
-  var fileMappings = [
+function getFileMappingsFromName(name) {
+  return [
     {
       oldPath: path.join('app', 'components', name + '.js'),
       newPath: path.join('addon', name, 'component.js')
@@ -74,22 +29,20 @@ function overwriteFiles(name) {
       newPath: path.join('addon', name, 'style.css')
     }
   ];
-
-  return Promise.map(fileMappings, function(fileMapping) {
-    return copyFileIfExists(fileMapping.oldPath, fileMapping.newPath);
-  });
 }
 
 module.exports = function(name) {
   ui.start('Extracting component ' + name + ' into addon/' + name + '\n');
 
-  return createAddonFolder(name).then(function() {
-    return copyBlueprintFiles(name, 'micro-component');
-  }).then(function() {
-    return insertAddonName(PLACEHOLDER, name);
-  }).then(function () {
-    return overwriteFiles(name);
-  }).then(function() {
+  var options = {
+    addonName: name,
+    blueprintName: 'micro-component',
+    placeholder: '<%= componentName %>',
+    filesToReplacePlaceholderIn: getFilesWithPlaceholdersFromName(name),
+    fileMappings: getFileMappingsFromName(name)
+  }
+
+  return extractAddon(options).then(function() {
     ui.write('Succesfully extracted component\n');
     return Promise.resolve();
   });
